--- design-patterns/src/main/java/com/huajieyu/pricinples/demo04/interfaceIsolation/after/Waterproof.java
@@ -1,5 +1,0 @@
-package com.huajieyu.pricinples.demo04.interfaceIsolation.after;
-
-public interface Waterproof {
-    void waterproof();
-}
