+++ design-patterns/src/main/java/com/huajieyu/pricinples/demo4/interfaceIsolation/after/AdvancedSafetyTest.java
@@ -1,4 +1,4 @@
-package com.huajieyu.pricinples.demo04.interfaceIsolation.after;
+package com.huajieyu.pricinples.demo4.interfaceIsolation.after;
 
 import org.junit.jupiter.api.Test;
