+++ design-patterns/src/main/java/com/huajieyu/pricinples/demo4/interfaceIsolation/after/Fireproof.java
@@ -1,0 +1,5 @@
+package com.huajieyu.pricinples.demo4.interfaceIsolation.after;
+
+public interface Fireproof {
+    void fireproof();
+}
