+++ design-patterns/src/main/java/com/huajieyu/pricinples/demo4/interfaceIsolation/after/MySafetyDoor.java
@@ -1,4 +1,4 @@
-package com.huajieyu.pricinples.demo04.interfaceIsolation.after;
+package com.huajieyu.pricinples.demo4.interfaceIsolation.after;
 
 public class MySafetyDoor implements AntiTheft, Fireproof, Waterproof {
     @Override
