+++ design-patterns/src/main/java/com/huajieyu/pricinples/demo4/interfaceIsolation/after/Waterproof.java
@@ -1,0 +1,5 @@
+package com.huajieyu.pricinples.demo4.interfaceIsolation.after;
+
+public interface Waterproof {
+    void waterproof();
+}
